Lua example warning removal fixes for vc++
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk@10539 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Examples/lua/arrays/example.c
@@ -16,7 +16,7 @@ void sort_int(int* arr, int len)
 // ditto doubles
 int compare_double(const void * a, const void * b)
 {
-  return ( *(double*)a - *(double*)b );
+  return (int)( *(double*)a - *(double*)b );
 }
 
 void sort_double(double* arr, int len)
